create package "tombstones"
